usb_queue: Call transfer completion callback while flushing
This allows the user to clean up any resources which the transfer held.
That being said, simply claiming zero transferred bytes isn't terribly
descriptive.

This is especially problematic in cases where you might want to schedule a
follow-up transfer in the completion callback. In the case the endpoint
is disabled by the time the user attempts to reschedule, leading to a
deadlock.